+++ src/m365/spo/commands.ts
@@ -137,6 +137,7 @@ export default {
   LISTITEM_RECORD_DECLARE: `${prefix} listitem record declare`,
   LISTITEM_RECORD_UNDECLARE: `${prefix} listitem record undeclare`,
   LISTITEM_REMOVE: `${prefix} listitem remove`,
+  LISTITEM_ROLEASSIGNMENT_ADD: `${prefix} listitem roleassignment add`,
   LISTITEM_ROLEASSIGNMENT_REMOVE: `${prefix} listitem roleassignment remove`,
   LISTITEM_ROLEINHERITANCE_BREAK: `${prefix} listitem roleinheritance break`,
   LISTITEM_ROLEINHERITANCE_RESET: `${prefix} listitem roleinheritance reset`,
@@ -283,6 +284,7 @@ export default {
   WEB_REMOVE: `${prefix} web remove`,
   WEB_ROLEASSIGNMENT_REMOVE: `${prefix} web roleassignment remove`,
   WEB_ROLEASSIGNMENT_ADD: `${prefix} web roleassignment add`,
+  WEB_ROLEINHERITANCE_BREAK: `${prefix} web roleinheritance break`,
   WEB_ROLEINHERITANCE_RESET: `${prefix} web roleinheritance reset`,
   WEB_SET: `${prefix} web set`
 };